move commands from constants.ts to @types/commands.ts
